third_party/WebKit/LayoutTests/inspector/console/console-format-collections-expected.txt

CONSOLE MESSAGE: line 13: [object HTMLCollection]
CONSOLE MESSAGE: line 17: [object HTMLCollection]
CONSOLE MESSAGE: line 21: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 25: [object HTMLAllCollection]
CONSOLE MESSAGE: line 29: [object HTMLFormControlsCollection]
CONSOLE MESSAGE: line 33: [object RadioNodeList]
CONSOLE MESSAGE: line 39: 1,2,
CONSOLE MESSAGE: line 42: [object Object]
CONSOLE MESSAGE: line 48: [object Arguments]
CONSOLE MESSAGE: line 52: [object DOMTokenList]
CONSOLE MESSAGE: line 55: [object Object]
CONSOLE MESSAGE: line 56: [object Object]
CONSOLE MESSAGE: line 58: [object Object]
CONSOLE MESSAGE: line 59: [object Object]
CONSOLE MESSAGE: line 60: [object Object]
CONSOLE MESSAGE: line 61: [object Object]
CONSOLE MESSAGE: line 62: [object Object]
CONSOLE MESSAGE: line 63: [object Object]
CONSOLE MESSAGE: line 13: [object HTMLCollection]
CONSOLE MESSAGE: line 17: [object HTMLCollection]
CONSOLE MESSAGE: line 21: [object HTMLOptionsCollection]
CONSOLE MESSAGE: line 25: [object HTMLAllCollection]
CONSOLE MESSAGE: line 29: [object HTMLFormControlsCollection]
CONSOLE MESSAGE: line 33: [object RadioNodeList]
CONSOLE MESSAGE: line 39: 1,2,
CONSOLE MESSAGE: line 42: [object Object]
CONSOLE MESSAGE: line 48: [object Arguments]
CONSOLE MESSAGE: line 52: [object DOMTokenList]
CONSOLE MESSAGE: line 55: [object Object]
CONSOLE MESSAGE: line 56: [object Object]
CONSOLE MESSAGE: line 58: [object Object]
CONSOLE MESSAGE: line 59: [object Object]
CONSOLE MESSAGE: line 60: [object Object]
CONSOLE MESSAGE: line 61: [object Object]
CONSOLE MESSAGE: line 62: [object Object]
CONSOLE MESSAGE: line 63: [object Object]
Tests that console nicely formats HTML Collections, NodeLists and DOMTokenLists.

console-format-collections.html:13 HTMLCollection(1)
console-format-collections.html:17 HTMLCollection(3)
console-format-collections.html:21 HTMLOptionsCollection(2)
console-format-collections.html:25 HTMLAllCollection(14)
console-format-collections.html:29 HTMLFormControlsCollection(3)
console-format-collections.html:33 RadioNodeList(2)
console-format-collections.html:39 Array(2)
console-format-collections.html:42 NonArrayWithLength
console-format-collections.html:48 Arguments(2)
console-format-collections.html:52 DOMTokenList(3)
console-format-collections.html:55 ArrayLike(5)
console-format-collections.html:56 ArrayLike(4294967295)
console-format-collections.html:58 ArrayLike
console-format-collections.html:59 ArrayLike
console-format-collections.html:60 ArrayLike
console-format-collections.html:61 ArrayLike
console-format-collections.html:62 ArrayLike
console-format-collections.html:63 ArrayLike
console-format-collections.html:13 [select#sel, sel: select#sel]
console-format-collections.html:17 (3) [script, script, script]
console-format-collections.html:21 (2) [option, option, selectedIndex: 0]
console-format-collections.html:25 (14) [html, head, script, script, script, body, p, div.c1.c2.c3, form#f, select#sel, option, option, input, input, f: form#f, sel: select#sel, x: NodeList(2)]
console-format-collections.html:29 (3) [select#sel, input, input, sel: select#sel, x: RadioNodeList(2)]
console-format-collections.html:33 (2) [input, input, value: ""]
console-format-collections.html:39 (2) [1, Array(2)]
console-format-collections.html:42 NonArrayWithLength {keys: Array(0)}
console-format-collections.html:48 (2) [1, "2", callee: ƒ, Symbol(Symbol.iterator): ƒ]
console-format-collections.html:52 (3) ["c1", "c2", "c3", value: "c1 c2 c3"]
console-format-collections.html:55 (5) [empty × 5]
console-format-collections.html:56 (4294967295) [empty × 4294967295]
console-format-collections.html:58 ArrayLike {length: -5}
console-format-collections.html:59 ArrayLike {length: 5.6}
console-format-collections.html:60 ArrayLike {length: NaN}
console-format-collections.html:61 ArrayLike {length: Infinity}
console-format-collections.html:62 ArrayLike {length: -0}
console-format-collections.html:63 ArrayLike {length: 4294967296}
